requirements.txt
flask==2.3.3
psycopg2-binary==2.9.7
selenium==4.15.2
pandas==2.1.1
beautifulsoup4==4.12.2
requests==2.31.0
openpyxl==3.1.2
lxml==4.9.3
webdriver-manager==4.0.1
schedule==1.2.0
python-dotenv==1.0.0
gunicorn==21.2.0